requirements.txt
livekit-agents>=0.8.0
livekit-plugins-google>=0.8.0
aiohttp>=3.8.0